+++ examples/read_file.py
@@ -1,6 +1,0 @@
-import mclang
-
-with open('output.lang', 'r') as r:
-    doc = mclang.load(r)
-    res = doc.tl('test')
-    print(f'"{res}"')