+++ examples/write.py
@@ -1,9 +1,0 @@
-import mclang
-
-obj = {
-    'test': 'This is cool!',
-    'test2': 'It worked!',
-}
-
-result = mclang.dumps(obj)
-print(result)