+++ CHANGELOG.md
@@ -4,5 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - unreleased
+## [1.1.0] - 8/27/2025
+
+### General
+
+- Added `__all__`
+- mclang will now correctly read comments and apply it to Lang.
+- Added translator to translate your lang file.
+- `.lang` files now use `utf-8` encoding.
+
+## [1.0.0] - 7/3/2023
+
 - Initial release